nrf_security: Corrected handling of multiple backends and glue layer
This commit ensures that glue layer is now correctly compiled when any
cipher / curve is selected in two backends.
Previously the glue layer was only compiled if AES ciphers were
selected in multiple backends.

This commit also corrects the situation where two backends are selected but glue layer is not needed.
This would result in platform_alt.h not beeing in include path.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>